# 🚀 Performance Optimizations: 5.9x Immediate Speedup + Path to 240x Improvement