… optimized_data_structures.py
… performance_optimizations.py
… performance_summary.md
… src/db/models.py
… src/matcher.py
… src/parser.py
… tests/test_cli.py
… tests/test_matcher.py
… tests/test_name_matcher.py